feat: Update fastlane button colors and conveyor icon
This commit updates the fill color of the buttons in the fastlane metadata from a brown shade (`#8e4e31`) to a blue shade (`#07689A`).

Additionally, the conveyor configuration is updated to use `template/logo.svg` as the application icon instead of a generated icon with a specific color gradient. The custom icon gradient configuration is commented out.

Finally, the commit includes a new icon image (`fastlane/metadata/android/en-US/images/icon.png`).