refactor: comment out unused wasmJs and nonJs configurations
This commit comments out the `wasmJs` and `nonJs` configurations within the `app/build.gradle.kts` file.

The `wasmJs` configuration, including its browser settings and webpack configuration, is commented out as it is not currently being used. Similarly, the `nonJsMain` source set and its associated dependency on `libs.kermit.koin` are also commented out.

Additionally, the `applyDefaultHierarchyTemplate` block, which includes the `nonJs` group, is commented out. The dependency on `libs.kermit.koin` within the common source set is retained as it is used elsewhere.