Refactor: Rename INetworkRepository to IMediaRepository
This commit renames the `INetworkRepository` interface and its fake implementation to `IMediaRepository` and `FakeNetworkRepository` (which now implements `IMediaRepository`) respectively.

The `FakeNetworkRepository` is updated to implement the `getAllMedia` function from the `IMediaRepository` interface, though it is currently a `TODO`. The previous `get` and `gotoGoogle` methods have been removed.

The `testDataModule` in `TestDataModule.kt` is updated to reflect this change, binding `FakeNetworkRepository` to `IMediaRepository::class`.